refactoring towards being able to start a new chain with parity
